requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2024.12.14
click==8.1.8
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.7
httpx==0.25.1
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.24.0
